refactor: rename path to om-transportportal
