+++ apps/data-norge/public/content/om-transportportal/nyheter/nyheter.nb.mdx
