+++ apps/data-norge/public/content/om-transportportal/roller-og-ansvar/roller-og-ansvar.en.mdx
