+++ apps/data-norge/src/app/[lang]/om-transportportal/[...slug]/page.tsx
@@ -1,5 +1,5 @@
 import { LocaleCodes } from "@fdk-frontend/localization";
-import DocsPage, { generateMetadata as docsPageMetadata } from "../../components/docs/docs-page";
+import DocsPage, { generateMetadata as docsPageMetadata } from "../../../components/docs/docs-page";
 
 interface Props {
   params: Promise<{
@@ -8,13 +8,13 @@ interface Props {
 }
 
 export const generateMetadata = async (props: Props) => {
-  return await docsPageMetadata({ rootContentDirectory: "transportportal", ...props });
+  return await docsPageMetadata({ rootContentDirectory: "om-transportportal", ...props });
 };
 
 export default async function Page(props: Props) {
   return (
     <DocsPage
-      rootContentDirectory="transportportal"
+      rootContentDirectory="om-transportportal"
       {...props}
     />
   );
